requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==2.0.1
Flask-Cors==3.0.10
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.10
Werkzeug==1.0.1
numpy==1.20.2
pandas==1.2.4
scikit-learn==0.24.2
cassandra-driver==3.25.0
xgboost==1.4.2


